refactor(equalizer): align service conventions with maintainer style
