+++ Tests/KasetTests/EqualizerServiceTests.swift
@@ -308,42 +308,3 @@ struct EqualizerServiceTests {
         #expect(mock.startCallCount == baseline + 1)
     }
 }
-
-// MARK: - MockEqualizerAudioEngine
-
-private final class MockEqualizerAudioEngine: EqualizerAudioEngineProtocol, @unchecked Sendable {
-    var startResult: Result<Void, EqualizerAudioEngine.StartFailure>
-    private(set) var startCallCount = 0
-    private(set) var stopCallCount = 0
-    private(set) var lastAppliedSettings: EQSettings?
-    private var running = false
-
-    /// Tests can flip this to simulate a tap that succeeds at start time
-    /// but never delivers audio (the silence-detection path).
-    var hasObservedAudio: Bool = true
-
-    init(startResult: Result<Void, EqualizerAudioEngine.StartFailure> = .success(())) {
-        self.startResult = startResult
-    }
-
-    var isRunning: Bool {
-        self.running
-    }
-
-    func start() -> Result<Void, EqualizerAudioEngine.StartFailure> {
-        self.startCallCount += 1
-        if case .success = self.startResult {
-            self.running = true
-        }
-        return self.startResult
-    }
-
-    func stop() {
-        self.stopCallCount += 1
-        self.running = false
-    }
-
-    func apply(settings: EQSettings) {
-        self.lastAppliedSettings = settings
-    }
-}
